watchdog: Tighten dependencies on WDT_DAVINCI
The WDT_DAVINCI driver is not safe to compile on 64bit platforms such as
allyesconfig on a 64bit host. Tighten the dependencies here to the
platforms which could use it today.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
Note that today no platforms enable it, but based on the compatible
string a number of platforms could, so I'm cc'ing platform maintainers
that may want to enable it.

Cc: Andrew Davis <afd@ti.com>
Cc: David Lechner <david@lechnology.com>
Cc: Adam Ford <aford173@gmail.com>