video: Add SPL_VIDEO_SYNC_CYCLIC_MS symbol
When compiling video-uclass.c for SPL and with SPL_CYCLIC enabled, we
also need to know SPL_VIDEO_SYNC_CYCLIC_MS as a value. This is most
easily resolved by adding a symbol to set this to the value of
VIDEO_SYNC_CYCLIC_MS as one shouldn't ever need these to differ.

Signed-off-by: Tom Rini <trini@konsulko.com>