dts: Make xPL_OF_PLATDATA depend on !COMPILE_TEST
When doing a COMPILE_TEST build it makes more sense to disable this
feature and test the common path than it does to test the uncommon path
of [ST]PL_OF_PLATDATA being enabled.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
I thought about this a bit, and one thing is that we do uncover a number
of drivers today which have partial OF_PLATDATA support but we have even
more drivers which assume OF_PLATADATA is disabled and then fail to
compile in xPL. This choice will give us the most overall coverage I
believe.