malloc: Add some function prototypes for USE_HOSTCC
With the changes in commit 8fbcc0e0e839 ("boot: Assure FDT is always at
8-byte aligned address") to call memalign(...) we need to have this
function prototype available in both host tools and U-Boot itself. The
way that include/malloc.h (which is found for <malloc.h> for both host
tools and U-Boot) works on Linux hosts. It works on how we test Windows
hosts. It however fails on macOS as include/linux/string.h has problems
with the compiler there. To work around this, modify include/malloc.h to
check for USE_HOSTCC and when true not use <linux/string.h> but instead
provide the prototype for memalign.

Fixes: 8fbcc0e0e839 ("boot: Assure FDT is always at 8-byte aligned address")
Signed-off-by: Tom Rini <trini@konsulko.com>